refactor: fix file name cases (#37)
