+++ tests/unit/store/getters.spec.ts
@@ -1,7 +1,7 @@
 import getters from '@/store/getters'
 import emptyState from '@/store/state'
 import Getters from '@/types/Getters'
-import ApplicationState from '@/types/applicationState'
+import ApplicationState from '@/types/ApplicationState'
 import Variant from '@/types/Variant'
 import Assessment from '@/types/Assessment'
 import Variable from '@/types/Variable'
