+++ tests/unit/views/SidebarView.spec.ts
@@ -2,7 +2,7 @@ import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils'
 import SidebarView from '@/views/SidebarView.vue'
 import Vue from 'vue'
 import Vuex from 'vuex'
-import ApplicationState from '@/types/applicationState'
+import ApplicationState from '@/types/ApplicationState'
 
 const localVue = createLocalVue()
 localVue.use(Vuex)
