+++ tests/unit/views/TreeView.spec.ts
@@ -2,7 +2,7 @@ import { mount, createLocalVue, Wrapper } from '@vue/test-utils'
 import TreeView from '@/views/TreeView.vue'
 import Vue from 'vue'
 import Vuex from 'vuex'
-import ApplicationState from '@/types/applicationState'
+import ApplicationState from '@/types/ApplicationState'
 
 const localVue = createLocalVue()
 localVue.use(Vuex)
